requirements.txt
# MaxMahon — runtime dependencies
# Install: py -m pip install -r requirements.txt
# Note: `pyjwt[crypto]` extra installs `cryptography` for ES256 JWT verification.

fastapi>=0.110
uvicorn[standard]>=0.27
sse-starlette>=2.0
pydantic>=2.5
python-dotenv>=1.0
apscheduler>=3.10
markdown>=3.5
requests>=2.31
httpx>=0.27
pandas>=2.1
anthropic>=0.34
thaifin>=0.6
yahooquery>=2.3
pyjwt[crypto]>=2.8.0
